+++ pkgs/firehose/lib/firehose.dart
@@ -22,6 +22,8 @@ const String _botSuffix = '[bot]';
 const String _githubActionsUser = 'github-actions[bot]';
 
 const String _publishBotTag = '## Package publishing';
+const String _publishBotDescription = 'If you have publishing permissions, '
+    'you can use the links below to publish the changes after merging this PR.';
 
 const String _ignoreWarningsLabel = 'publish-ignore-warnings';
 
@@ -73,7 +75,9 @@ Documentation at https://github.com/dart-lang/ecosystem/wiki/Publishing-automati
     );
 
     if (results.hasSuccess) {
-      final commentText = '$_publishBotTag\n\n$markdownTable';
+      final commentText = '$_publishBotTag\n\n'
+          '$_publishBotDescription\n\n'
+          '$markdownTable';
 
       if (existingCommentId != null) {
         final idFile = File('./output/commentId');
@@ -164,8 +168,12 @@ Saving existing comment id $existingCommentId to file ${idFile.path}''');
           github.notice(message: message);
           results.addResult(Result.fail(package, message));
         } else {
-          final result = Result.success(package, '**ready to publish**',
-              repoTag, repo.calculateReleaseUri(package, github));
+          final result = Result.success(
+            package,
+            '**ready to publish**',
+            repoTag,
+            repo.calculateReleaseUri(package, github),
+          );
           print(result);
           results.addResult(result);
         }
@@ -251,13 +259,16 @@ Saving existing comment id $existingCommentId to file ${idFile.path}''');
 
     if (pubspecVersion != tag.version) {
       stderr.writeln(
-          "Pubspec version ($pubspecVersion) and git tag ($tag) don't agree.");
+        "Pubspec version ($pubspecVersion) and git tag ($tag) don't agree.",
+      );
       return false;
     }
 
     if (pubspecVersion != changelogVersion) {
-      stderr.writeln('Pubspec version ($pubspecVersion) and changelog version '
-          "($changelogVersion) don't agree.");
+      stderr.writeln(
+        'Pubspec version ($pubspecVersion) and changelog version '
+        "($changelogVersion) don't agree.",
+      );
       return false;
     }
 
@@ -284,12 +295,7 @@ Saving existing comment id $existingCommentId to file ${idFile.path}''');
     }
     return await runCommand(
       command,
-      args: [
-        'pub',
-        'publish',
-        if (dryRun) '--dry-run',
-        if (force) '--force',
-      ],
+      args: ['pub', 'publish', if (dryRun) '--dry-run', if (force) '--force'],
       cwd: package.directory,
     );
   }
@@ -331,8 +337,13 @@ class Result {
   final String? gitTag;
   final Uri? publishReleaseUri;
 
-  Result(this.severity, this.package, this.message,
-      [this.gitTag, this.publishReleaseUri]);
+  Result(
+    this.severity,
+    this.package,
+    this.message, [
+    this.gitTag,
+    this.publishReleaseUri,
+  ]);
 
   factory Result.fail(Package package, String message) =>
       Result(Severity.error, package, message);
@@ -340,8 +351,12 @@ class Result {
   factory Result.info(Package package, String message) =>
       Result(Severity.info, package, message);
 
-  factory Result.success(Package package, String message,
-          [String? gitTag, Uri? publishReleaseUri]) =>
+  factory Result.success(
+    Package package,
+    String message, [
+    String? gitTag,
+    Uri? publishReleaseUri,
+  ]) =>
       Result(Severity.success, package, message, gitTag, publishReleaseUri);
 
   @override